Fix check for existing output html file
The problem is, there is no checking if the output html file is already
exist or not. So then, when executing the same operation again on the
same text file, the existing html file will be opened again to append
the result of the processed text file.

To fix this issue, implement a file check of the output html file if it
already exist or not.